+++ NEWS.md
@@ -1,6 +1,7 @@
 # emvoGCMS 0.6.2
 
 - add getMSP to read in EI-MS or MS/MS msp file as list for annotation
+- add xrankanno and dotpanno function for MS/MS X Rank and dot product annotation
 
 # enviGCMS 0.6.1